src/requirements.txt
pip
RPi.GPIO
